fix(compile): fix compiling error of "comm::Domain::Builder is not a template type" in hipclang env
By using hipcc as compiler(with clang-12), if we include header files of this lib in another
lib/binary, the compiling step may cause an error "error: 'comm::Domain::Builder' is not a template
type. template<typename, typename>". We can fix this problem by adjust the position of declaration
of `class Builder;`.